Renamed files for inclusive language
