+++ Tryouts/Plugins/ApplicationPlugins/process explorer/src/app/main-view/connections/connections.component.html
