+++ Tryouts/Plugins/ApplicationPlugins/process explorer/src/app/main-view/main-view-routing.module.ts
@@ -1,7 +1,7 @@
 /* Morgan Stanley makes this available to you under the Apache License, Version 2.0 (the "License"). You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0. See the NOTICE file distributed with this work for additional information regarding copyright ownership. Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License. */
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { MasterViewComponent } from './master-view.component';
+import { MainViewComponent } from './main-view.component';
 import { ConnectionsComponent } from './connections/connections.component';
 import { ProcessesComponent } from './processes/processes.component';
 //import { MemoryChartComponent } from './memory-chart/memory-chart.component';
@@ -10,7 +10,7 @@ import { SubsystemComponent } from './subsystems/subsystem.component';
 const routes: Routes = [
   { 
     path: '', 
-    component: MasterViewComponent, 
+    component: MainViewComponent, 
     children: [
       { 
         path: '', 
@@ -38,5 +38,5 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class MasterViewRoutingModule {
+export class MainViewRoutingModule {
 }
