+++ Tryouts/Plugins/ApplicationPlugins/process explorer/src/app/main-view/main-view.component.html
@@ -12,15 +12,15 @@
       <span [isHeader]="true" igxDrawerItem>
         <span class="navigation-drawer_1">Views in this app</span>
       </span>
-      <span [isHeader]="false" igxDrawerItem routerLink="/master-view/processes">
+      <span [isHeader]="false" igxDrawerItem routerLink="/main-view/processes">
         <igx-icon font="material" [active]="true">home</igx-icon>
         <span class="navigation-drawer_1">Processes</span>
       </span>
-      <span [isHeader]="false" igxDrawerItem routerLink="/master-view/connections">
+      <span [isHeader]="false" igxDrawerItem routerLink="/main-view/connections">
         <igx-icon font="material" [active]="true">dns</igx-icon>
         <span class="navigation-drawer_1">Connections</span>
       </span>
-      <span [isHeader]="false" igxDrawerItem routerLink="/master-view/subsystems">
+      <span [isHeader]="false" igxDrawerItem routerLink="/main-view/subsystems">
         <igx-icon font="material" [active]="true">dns</igx-icon>
         <span class="navigation-drawer_1">Subsystems</span>
       </span>
