+++ Tryouts/Plugins/ApplicationPlugins/process explorer/src/app/main-view/main-view.component.scss
