+++ Tryouts/Plugins/ApplicationPlugins/process explorer/src/app/main-view/main-view.component.spec.ts
@@ -5,22 +5,22 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { IgxNavbarModule, IgxButtonModule, IgxToggleModule, IgxIconModule, IgxNavigationDrawerModule, IgxListModule } from 'igniteui-angular';
-import { MasterViewComponent } from './master-view.component';
+import { MainViewComponent } from './main-view.component';
 
-describe('MasterViewComponent', () => {
-  let component: MasterViewComponent;
-  let fixture: ComponentFixture<MasterViewComponent>;
+describe('MainViewComponent', () => {
+  let component: MainViewComponent;
+  let fixture: ComponentFixture<MainViewComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ MasterViewComponent ],
+      declarations: [ MainViewComponent ],
       imports: [ NoopAnimationsModule, FormsModule, RouterTestingModule, IgxNavbarModule, IgxButtonModule, IgxToggleModule, IgxIconModule, IgxNavigationDrawerModule, IgxListModule ]
     })
     .compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(MasterViewComponent);
+    fixture = TestBed.createComponent(MainViewComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
