+++ Tryouts/Plugins/ApplicationPlugins/process explorer/src/app/main-view/processes/ServiceProcessObject.ts
