+++ Tryouts/Plugins/ApplicationPlugins/process explorer/src/app/main-view/processes/processes.component.html
@@ -1,6 +1,6 @@
 <!-- Morgan Stanley makes this available to you under the Apache License, Version 2.0 (the "License"). You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0. See the NOTICE file distributed with this work for additional information regarding copyright ownership. Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License. -->
 <div class="column-layout column">
-  <button igxButton="icon" type="icon" [disabled]="false" routerLink="/master-view/memory-chart">
+  <button igxButton="icon" type="icon" [disabled]="false" routerLink="/main-view/memory-chart">
     <igx-icon>open_in_new</igx-icon>
   </button>
 </div>
