+++ Tryouts/Plugins/ApplicationPlugins/process explorer/src/app/main-view/subsystems/ServiceSubsystemObject.ts
